# app/marksman/internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/notification-groups:
        get:
            tags:
                - NotificationGroup
            operationId: NotificationGroup_ListNotificationGroup
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.ListNotificationGroupReply'
        post:
            tags:
                - NotificationGroup
            operationId: NotificationGroup_CreateNotificationGroup
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.CreateNotificationGroupRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.CreateNotificationGroupReply'
    /v1/notification-groups/{notificationGroupUid}/subscription:
        get:
            tags:
                - NotificationGroupSubscription
            operationId: NotificationGroupSubscription_GetNotificationGroupSubscription
            parameters:
                - name: notificationGroupUid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.GetNotificationGroupSubscriptionReply'
        put:
            tags:
                - NotificationGroupSubscription
            operationId: NotificationGroupSubscription_SaveNotificationGroupSubscription
            parameters:
                - name: notificationGroupUid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.SaveNotificationGroupSubscriptionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.SaveNotificationGroupSubscriptionReply'
    /v1/notification-groups/{uid}:
        get:
            tags:
                - NotificationGroup
            operationId: NotificationGroup_GetNotificationGroup
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.NotificationGroupItem'
        put:
            tags:
                - NotificationGroup
            operationId: NotificationGroup_UpdateNotificationGroup
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.UpdateNotificationGroupRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.UpdateNotificationGroupReply'
        delete:
            tags:
                - NotificationGroup
            operationId: NotificationGroup_DeleteNotificationGroup
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.DeleteNotificationGroupReply'
...
components:
    schemas:
...
        marksman.api.v1.CreateNotificationGroupReply:
            type: object
            properties:
                uid:
                    type: string
        marksman.api.v1.CreateNotificationGroupRequest:
            type: object
            properties:
                name:
                    type: string
                remark:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                members:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.NotificationMemberRequest'
                webhooks:
                    type: array
                    items:
                        type: string
                templates:
                    type: array
                    items:
                        type: string
                emailConfigs:
                    type: array
                    items:
                        type: string
...
        marksman.api.v1.DeleteNotificationGroupReply:
            type: object
            properties: {}
...
        marksman.api.v1.GetNotificationGroupSubscriptionReply:
            type: object
            properties:
                filter:
                    $ref: '#/components/schemas/marksman.api.v1.SubscriptionFilter'
...
        marksman.api.v1.ListNotificationGroupReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.NotificationGroupItem'
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
...
        marksman.api.v1.NotificationGroupItem:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                remark:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                status:
                    type: integer
                    format: enum
                createdAt:
                    type: string
                updatedAt:
                    type: string
                webhooks:
                    type: array
                    items:
                        type: string
                templates:
                    type: array
                    items:
                        type: string
                emailConfigs:
                    type: array
                    items:
                        type: string
                members:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.NotificationMemberItem'
                webhookItems:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.NotificationResourceItem'
                templateItems:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.NotificationResourceItem'
                emailConfigItems:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.NotificationResourceItem'
        marksman.api.v1.NotificationMemberItem:
            type: object
            properties:
                memberUid:
                    type: string
                isEmail:
                    type: boolean
                isPhone:
                    type: boolean
                memberName:
                    type: string
                memberAvatar:
                    type: string
        marksman.api.v1.NotificationMemberRequest:
            type: object
            properties:
                memberUid:
                    type: string
                isEmail:
                    type: boolean
                isPhone:
                    type: boolean
        marksman.api.v1.NotificationResourceItem:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
...
        marksman.api.v1.SaveNotificationGroupSubscriptionReply:
            type: object
            properties: {}
        marksman.api.v1.SaveNotificationGroupSubscriptionRequest:
            type: object
            properties:
                notificationGroupUid:
                    type: string
                filter:
                    $ref: '#/components/schemas/marksman.api.v1.SubscriptionFilter'
...
        marksman.api.v1.StrategyLevelPair:
            type: object
            properties:
                strategyUid:
                    type: string
                levelUid:
                    type: string
...
        marksman.api.v1.SubscriptionFilter:
            type: object
            properties:
                strategyGroupUids:
                    type: array
                    items:
                        type: string
                strategyUids:
                    type: array
                    items:
                        type: string
                strategyLevels:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.StrategyLevelPair'
                datasourceUids:
                    type: array
                    items:
                        type: string
                labels:
                    type: object
                    additionalProperties:
                        type: string
                excludeLabels:
                    type: object
                    additionalProperties:
                        type: string
                datasourceLevelUids:
                    type: array
                    items:
                        type: string
            description: |-
                Subscription filter: when multiple dimensions are set, an alert matches if it satisfies at least one (OR).
                 - strategyGroupUids: alert's strategy must belong to one of these groups.
                 - strategyUids: alert's strategy_uid must be in this list.
                 - strategyLevels: alert's (strategy_uid, level_uid) must be in this list.
                 - labels: alert must contain all these key-value pairs with the same values.
                 - excludeLabels: alert must not contain any of these key-value pairs (subset match).
                 - datasourceUids: alert's datasource must be in this list.
                 - datasourceLevelUids: alert datasource's level_uid must be in this list.
...
        marksman.api.v1.UpdateNotificationGroupReply:
            type: object
            properties: {}
        marksman.api.v1.UpdateNotificationGroupRequest:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                remark:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                members:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.NotificationMemberRequest'
                webhooks:
                    type: array
                    items:
                        type: string
                templates:
                    type: array
                    items:
                        type: string
                emailConfigs:
                    type: array
                    items:
                        type: string
...