# app/marksman/pkg/api/v1/notification_group_subscription_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.29.2
// source: marksman/api/v1/notification_group_subscription.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationNotificationGroupSubscriptionGetNotificationGroupSubscription = "/marksman.api.v1.NotificationGroupSubscription/GetNotificationGroupSubscription"
const OperationNotificationGroupSubscriptionSaveNotificationGroupSubscription = "/marksman.api.v1.NotificationGroupSubscription/SaveNotificationGroupSubscription"

type NotificationGroupSubscriptionHTTPServer interface {
	GetNotificationGroupSubscription(context.Context, *GetNotificationGroupSubscriptionRequest) (*GetNotificationGroupSubscriptionReply, error)
	SaveNotificationGroupSubscription(context.Context, *SaveNotificationGroupSubscriptionRequest) (*SaveNotificationGroupSubscriptionReply, error)
}

func RegisterNotificationGroupSubscriptionHTTPServer(s *http.Server, srv NotificationGroupSubscriptionHTTPServer) {
	r := s.Route("/")
	r.GET("/v1/notification-groups/{notificationGroupUid}/subscription", _NotificationGroupSubscription_GetNotificationGroupSubscription0_HTTP_Handler(srv))
	r.PUT("/v1/notification-groups/{notificationGroupUid}/subscription", _NotificationGroupSubscription_SaveNotificationGroupSubscription0_HTTP_Handler(srv))
}

func _NotificationGroupSubscription_GetNotificationGroupSubscription0_HTTP_Handler(srv NotificationGroupSubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetNotificationGroupSubscriptionRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNotificationGroupSubscriptionGetNotificationGroupSubscription)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetNotificationGroupSubscription(ctx, req.(*GetNotificationGroupSubscriptionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetNotificationGroupSubscriptionReply)
		return ctx.Result(200, reply)
	}
}

func _NotificationGroupSubscription_SaveNotificationGroupSubscription0_HTTP_Handler(srv NotificationGroupSubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SaveNotificationGroupSubscriptionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNotificationGroupSubscriptionSaveNotificationGroupSubscription)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SaveNotificationGroupSubscription(ctx, req.(*SaveNotificationGroupSubscriptionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SaveNotificationGroupSubscriptionReply)
		return ctx.Result(200, reply)
	}
}

type NotificationGroupSubscriptionHTTPClient interface {
	GetNotificationGroupSubscription(ctx context.Context, req *GetNotificationGroupSubscriptionRequest, opts ...http.CallOption) (rsp *GetNotificationGroupSubscriptionReply, err error)
	SaveNotificationGroupSubscription(ctx context.Context, req *SaveNotificationGroupSubscriptionRequest, opts ...http.CallOption) (rsp *SaveNotificationGroupSubscriptionReply, err error)
}

type NotificationGroupSubscriptionHTTPClientImpl struct {
	cc *http.Client
}

func NewNotificationGroupSubscriptionHTTPClient(client *http.Client) NotificationGroupSubscriptionHTTPClient {
	return &NotificationGroupSubscriptionHTTPClientImpl{client}
}

func (c *NotificationGroupSubscriptionHTTPClientImpl) GetNotificationGroupSubscription(ctx context.Context, in *GetNotificationGroupSubscriptionRequest, opts ...http.CallOption) (*GetNotificationGroupSubscriptionReply, error) {
	var out GetNotificationGroupSubscriptionReply
	pattern := "/v1/notification-groups/{notificationGroupUid}/subscription"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationNotificationGroupSubscriptionGetNotificationGroupSubscription))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *NotificationGroupSubscriptionHTTPClientImpl) SaveNotificationGroupSubscription(ctx context.Context, in *SaveNotificationGroupSubscriptionRequest, opts ...http.CallOption) (*SaveNotificationGroupSubscriptionReply, error) {
	var out SaveNotificationGroupSubscriptionReply
	pattern := "/v1/notification-groups/{notificationGroupUid}/subscription"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNotificationGroupSubscriptionSaveNotificationGroupSubscription))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}